Rename file validator,js to validate.js
